refactorizing and update profile
