+++ src/css/home.css
@@ -60,7 +60,6 @@ grid-template-areas: ". pictureAside userName . "
       "content";
   }
 
-
 .mainHead {
     display: grid;
     grid-template-columns: 15% 70% 15%;
@@ -69,31 +68,32 @@ grid-template-areas: ". pictureAside userName . "
 }
 
 .userPicture {
-    min-width: 75px;
-    max-width: 50px;
     display: block;
     margin-left: auto;
     margin-right: auto;
-    padding-top: 10px;
-    padding-right: 20px;
-    padding-bottom: 10px;
-    padding-left: 20px;
     cursor: pointer;
+    object-fit: cover;
+    border-radius: 50%;
+    width: 50px;
+    height: 50px;
 }
 
 .userPictureAside {
-    text-align: center;
     min-width: 50px;
-    max-width: 100px;
+    max-width: 150px;
     grid-area: pictureAside;
-    display: block;
     margin-left: auto;
     margin-right: auto;
+    object-fit: cover;
+    display: block;
+    height: 150px;
+    margin-bottom: -10px;
+    border-radius: 250px;
 }
 .userName {
     grid-area: userName;
     align-self: center;
-    margin-top: 50px;
+    margin-top: 80px;
     font-weight: bold;
 }
