+++ src/firebase-controller/storage.js
@@ -1,4 +1,4 @@
-import { user } from './autentication.js';
+import { user } from '../firebase/autentication.js';
 
 export const uploadProfileImg = (file) => {
   const currentUser = user();
@@ -16,7 +16,7 @@ export const uploadProfileImg = (file) => {
       db.collection('users-qa').doc(currentUser.uid).update({
         photo: url,
       });
-      alert('Foto de perfil actualizada');
+      //   alert('Foto de perfil actualizada');
       const img = document.querySelector('.img-edit-user-profile');
       img.src = url;
     });
